Moved Sparkline and SparklineTests to Common element directories, minor JDIAction description/comment tweaks
