+++ jdi-light-vuetify/src/main/java/com/epam/jdi/light/vuetify/asserts/SparklineAssert.java
@@ -2,7 +2,7 @@ package com.epam.jdi.light.vuetify.asserts;
 
 import com.epam.jdi.light.asserts.generic.UIAssert;
 import com.epam.jdi.light.common.JDIAction;
-import com.epam.jdi.light.vuetify.elements.complex.Sparkline;
+import com.epam.jdi.light.vuetify.elements.common.Sparkline;
 import org.hamcrest.Matchers;
 
 import java.util.List;
