+++ jdi-light-vuetify/src/main/java/com/epam/jdi/light/vuetify/elements/common/Sparkline.java
@@ -1,4 +1,4 @@
-package com.epam.jdi.light.vuetify.elements.complex;
+package com.epam.jdi.light.vuetify.elements.common;
 
 import com.epam.jdi.light.asserts.generic.HasAssert;
 import com.epam.jdi.light.common.JDIAction;
@@ -19,6 +19,11 @@ import static java.lang.Double.parseDouble;
  */
 
 public class Sparkline extends UIBaseElement<SparklineAssert> implements HasAssert<SparklineAssert> {
+    // The Sparkline Vue element is an SVG image built around a single SVG <path> (line) sub-element
+
+    private UIElement getPath() {
+        return core().find("path");
+    }
 
     @JDIAction("Get '{name}' line width")
     public double getLineWidth() {
@@ -28,16 +33,12 @@ public class Sparkline extends UIBaseElement<SparklineAssert> implements HasAsse
     @JDIAction("Check if '{name}' is filled")
     public boolean isFilled() { return !getPath().getAttribute("fill").equals("none"); }
 
-    @JDIAction("Get '{name}' path sub-element")
-    private UIElement getPath() {
-        return core().find("path");
-    }
-
-    @JDIAction("Get value of the shape attribute of '{name}' path sub-element")
+    @JDIAction("Get '{name}' SVG-specific path shape definition")
     public String getPathShape() {
         return getPath().getAttribute("d");
     }
 
+    // returns an empty list if a sparkline has no data points
     @JDIAction("Get data point label texts of '{name}'")
     public List<String> getLabelTexts() {
         return core().finds("text").stream()
@@ -45,7 +46,7 @@ public class Sparkline extends UIBaseElement<SparklineAssert> implements HasAsse
                 .collect(Collectors.toList());
     }
 
-    @JDIAction("Get map representation (keys: offsets, values: stop-colors) of '{name}' linear gradient sub-element")
+    @JDIAction("Get map representation (keys: offsets, values: stop-colors) of '{name}' linear gradient")
     public Map<String, String> getLinearGradientMap() {
         Map<String, String> linearGradientMap = new HashMap<>();
         // core.finds() somehow fails to find elements with the following selector
